refactor: productcatalog http mappers
